Eliminate (accidental) progressive erosion.
Previous erosion is still used for Drainage.recompute, but we
again start from scratch in Waterflow.recompute.  As written before
the number of erosion/sedimentation cycles was a function of the
order and scope of topo/hydro editing operations.  Now they are 
simply a function of the topography, rainfall, and incoming rivers.